[#43] refactor: router 폴더 구조 변경
